population/requirements.txt
# Data population scripts (PostgreSQL + MongoDB)
psycopg[binary]>=3.1
pymongo>=4.6
python-dotenv>=1.0
